pyproject.toml
[tool.poetry]
name = "cmd-j-backend"
version = "0.1.0"
description = ""
authors = ["Nikita <hormold@gmail.com>"]
readme = "README.md"
packages = [{include = "cmd_j_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.26.5"
Flask = "^2.2.3"
Flask-Cors = "^3.0.10"
python-dotenv = "^0.21.1"
waitress = "^2.1.2"
duckduckgo-search = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
